+++ src/json-safe-parser.ts
@@ -1,6 +1,6 @@
 // vim: tabstop=8 softtabstop=0 noexpandtab shiftwidth=8 nosmarttab
 
-import { z, ZodIssueCode } from 'zod';
+import { z } from 'zod';
 
 function parseJsonPreprocessor(value: any, ctx: z.RefinementCtx) {
 	if(typeof value === 'string') {